refactor: split tau-tools registry/core module (closes #1749)
